Changed reporting of ph, temp, humidity from int to double; VERSION 1.1

+++ headers/Si7021.h
@@ -19,8 +19,8 @@ void Si7021_init(void);
 void Si7021_reset(void);
 void Si7021_read_humidity(void);
 void Si7021_read_temp(void);
-int Si7021_avg_humidity(void);
-int Si7021_avg_temp(void);
+double Si7021_avg_humidity(void);
+double Si7021_avg_temp(void);
 int Si7021_set_heater(char heat);
 #ifdef	__cplusplus
 extern "C" {
